Refactor wallpaper logic into WallpaperManager class
